+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### New features
 
-  - Add new `dalesbred-kotlin` module which adds Kotlin extensions to `Database`.
+  - Add optional Kotlin-support; `org.dalesbred.integration.kotlin` package adds Kotlin extensions to `Database`.
   
 ## 1.2.3 (2016-12-01)